Sidechain version fork height management
Modified the ContextualCheck() function to enforce that a sidechain creation transaction uses only the allowed version (before and after the Sidechain Version fork). This function is called both in AcceptTxToMemoryPool() and in AcceptBlock().

In order to implement some unit tests for all the scenarios, the test framework has been extended to call AcceptTxToMemoryPool() safely.